Fixed error, removed deprecated file structure, merged function with main python file.
